Relocate generator info structs
* Relocate `GeneratorInfo` and associated structs into
  `homotopy-graphics`.
* Rename instances of 'color' to 'style' where relevant.
