+++ homotopy-graphics/Cargo.toml
@@ -21,6 +21,7 @@ lyon_algorithms = "0.17.6"
 lyon_geom = "0.17.7"
 lyon_path = "0.17.7"
 paste = "1.0.7"
+palette = { git = "https://github.com/Ogeon/palette", features=["serializing"] }
 petgraph = "0.6.2"
 serde = { features = ["derive"], version = "1.0.137" }
 serde_json = "1.0.82"
